+++ apps/site/src/pages/Components/Table.tsx
@@ -5,7 +5,7 @@ import {
   boolean,
   select,
   signalTarget,
-} from '@chronista-club/creoui-editor-host'
+} from '@chronista-club/creo-ui-editor-host'
 import { A } from '@solidjs/router'
 import { createSignal } from 'solid-js'
 
@@ -214,7 +214,7 @@ export default function Table() {
         <div class="docs-playground-frame">
           <EditorHostProvider
             config={{
-              localStorageNamespace: 'creoui-docs.table-editor',
+              localStorageNamespace: 'creo-ui-docs.table-editor',
             }}
           >
             <TableEditorDemo />